fix: 상품 재고 필드 Long -> int 수정

+++ src/main/java/com/danahub/zipitda/store/dto/ProductDetailResponseDto.java
@@ -1,5 +1,6 @@
 package com.danahub.zipitda.store.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -9,8 +10,8 @@ public record ProductDetailResponseDto(
         String name,
         String description,
         String categoryName,
-        Long price,
-        Long stockQuantity,
+        BigDecimal price,
+        int stockQuantity,
         LocalDateTime createdAt,
         LocalDateTime updatedAt,
         List<String> imageUrls
